refactor(core): merge shared/ and infra/ into utils/
Consolidate shared kernel types and infrastructure utilities into a
single utils/ module. Rename DeviceContext to device::Context.
