--- core/src/infra.rs
@@ -1,3 +1,0 @@
-pub mod fs_helpers;
-pub mod markdown;
-pub mod paths;
