+++ core/src/lib.rs
@@ -1,18 +1,12 @@
-mod infra;
 mod note;
 mod project;
-pub mod shared;
 pub mod sync;
 mod timeline;
+pub mod utils;
 
 mod error;
 
 pub use error::CoreError;
-pub use infra::markdown::{format_note_markdown, format_timeline_line};
-pub use infra::paths::{
-    active_tasks_dir, done_tasks_dir, note_file_path, project_dir, project_file_path, projects_dir,
-    timeline_file_path,
-};
 pub use note::error::NoteError;
 pub use note::{
     NoteSummary, create_draft_note, delete_note, list_notes, read_note, read_note_by_filename,
@@ -24,8 +18,13 @@ pub use project::{
     create_task, delete_task, get_project_activity_summary, list_active_tasks, list_done_tasks,
     list_projects, read_project, update_task,
 };
-pub use shared::context::DeviceContext;
-pub use shared::frontmatter;
-pub use shared::validated::{Filename, NoteFilename, Slug};
 pub use timeline::error::TimelineError;
 pub use timeline::{list_timeline_dates, read_timeline, save_timeline_entry};
+pub use utils::device::Context as DeviceContext;
+pub use utils::frontmatter;
+pub use utils::markdown::{format_note_markdown, format_timeline_line};
+pub use utils::paths::{
+    active_tasks_dir, done_tasks_dir, note_file_path, project_dir, project_file_path, projects_dir,
+    timeline_file_path,
+};
+pub use utils::validated::{Filename, NoteFilename, Slug};
