--- core/src/shared.rs
@@ -1,3 +1,0 @@
-pub mod context;
-pub mod frontmatter;
-pub mod validated;
