+++ core/src/utils/validated.rs
